Drop the assistive flag; helper model is command-driven; trim latexFixer prompt
Per maintainer feedback, remove the `assistive` agent-definition flag entirely.
The helper model is now used purely when a "fix LaTeX" VS Code action opts in via
`preferHelperModel` (the Fix-Compilation command and the progress-view compile
fixer) — there is no registry flag.

- Remove `assistive` from the settings schemas (base + strict input), AgentEntry,
  the YAML scanner, and the latexFixer/latexDiff YAMLs.
- Replace assistiveModel.ts with helperModelPreference.ts: `preferHelperModel()`
  swaps to the configured helper model when a command opts in, guarding only on
  tool-capability (for tool-use agents) and availability — no entry lookup, so
  the source-pin/launch-resolver concern no longer applies.
- Rename the runAgent option `preferHelperModelForAssistive` → `preferHelperModel`
  and update the executeCommand/ProgressView wiring + the controller comment.
- Trim the latexFixer prompt substantially and make it explicit for both run
  modes: editing workspace files directly is the normal job; `/executions`
  output is inspect-only and editing the matching workspace file needs the user's
  confirmation.
- Replace/trim the tests (new helperModelPreference test; drop the assistive
  scanner test and registry regression guard).
