+++ src/agent/index/agentEntry.ts
@@ -18,7 +18,6 @@ export interface AgentEntry {
   rounds?: number; // workflow round count
   visibility?: string[]; // remote only: group names that can access the agent
   internal?: boolean; // internal agents are hidden from dropdowns but launchable by commands
-  assistive?: boolean; // support/helper agents prefer the helper model when launched by the fix-LaTeX actions
 }
 
 /**
