+++ src/agent/index/agentYamlScanner.ts
@@ -212,7 +212,6 @@ function scanYaml(
         : AgentCategory.Workflow;
 
     const internal = rawSettings.internal === true || undefined;
-    const assistive = rawSettings.assistive === true || undefined;
     let rounds: number | undefined;
     if (
       category === AgentCategory.Workflow &&
@@ -239,7 +238,6 @@ function scanYaml(
         : undefined,
       rounds,
       internal,
-      assistive,
     };
   } catch (err) {
     logger.warn(
